requirements.txt
chardet==7.4.0.post2
deepl==1.29.0
google_trans_new==1.1.9
langid==1.1.6
ollama==0.6.1
prettytable==3.17.0
Requests==2.33.1
tencentcloud_sdk_python_common==3.1.70
tencentcloud_sdk_python_tmt==3.1.70
torch==2.11.0+cu130
transformers==4.56.0
volcengine_python_sdk==5.0.21
xfyunsdknlp==0.0.3
